+++ source/Classes/DirectoryActivatedRoleCompleter.ps1
@@ -14,7 +14,7 @@ class DirectoryActivatedRoleCompleter : IArgumentCompleter {
         $ErrorActionPreference = 'Stop'
         try {
             Write-Progress -Id 51806 -Activity 'Get Activated Directory Roles' -Status 'Fetching from Azure' -PercentComplete 1
-            [List[CompletionResult]]$Result = Get-OPIMDirectoryRole -Activated | ForEach-Object {
+            [List[CompletionResult]]$Result = & ([scriptblock]::Create('Get-OPIMDirectoryRole -Activated')) | ForEach-Object {
                 $Scope = if ($PSItem.directoryScopeId -ne '/') {
                     "-> $($PSItem.directoryScope.displayName) "
                 }