+++ source/Classes/DirectoryEligibleRoleCompleter.ps1
@@ -14,7 +14,7 @@ class DirectoryEligibleRoleCompleter : IArgumentCompleter {
         $ErrorActionPreference = 'Stop'
         try {
             Write-Progress -Id 51806 -Activity 'Get Eligible Directory Roles' -Status 'Fetching from Azure' -PercentComplete 1
-            [List[CompletionResult]]$Result = Get-OPIMDirectoryRole | ForEach-Object {
+            [List[CompletionResult]]$Result = & ([scriptblock]::Create('Get-OPIMDirectoryRole')) | ForEach-Object {
                 $Scope = if ($PSItem.directoryScopeId -ne '/') {
                     "-> $($PSItem.directoryScope.displayName) "
                 }