+++ source/Classes/GroupEligibleCompleter.ps1
@@ -14,7 +14,7 @@ class GroupEligibleCompleter : IArgumentCompleter {
         $ErrorActionPreference = 'Stop'
         try {
             Write-Progress -Id 51806 -Activity 'Get Eligible PIM Groups' -Status 'Fetching from Azure' -PercentComplete 1
-            [List[CompletionResult]]$Result = Get-OPIMEntraIDGroup | ForEach-Object {
+            [List[CompletionResult]]$Result = & ([scriptblock]::Create('Get-OPIMEntraIDGroup')) | ForEach-Object {
                 "'{0} - {1} ({2})'" -f $PSItem.group.displayName, $PSItem.accessId, $PSItem.id
             } | Where-Object {
                 if (-not $wordToComplete) { return $true }